fix: Searching books would sometimes throw a warning
The warning was about the usage of setState on an unmounted component,
because fast searches would sometimes returned different lists of books
consecutively and the book component was executing a setState
instruction but some of them were not even being actually rendered,
because of new list of books being returned whenever the query string
changed.
Because of this issue I decided to opt for a totally different approach
for the type of state managed by App.js. As a consequence this lead to
a much simpler component for the Book.js, turned into a functional
component and no longer doing setState or AJAX request, but relying only
on the props received.